pyproject.toml
[project]
name = "agent-chat"
version = "0.1.0"
description = "This project is an AI Agent that can use models from both OpenAI (ChatGPT) and Ollama Providers to help understand AI Agents in a simple way, test various models, and display token usage."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
